add System abstract class make all systems to be child of System make land manager an observable
